pyproject.toml
[project]
name = "mq-trace-lite"
version = "0.1.0"
description = "CLI-first IBM MQ debugging and incident tool"
authors = [
    {name = "Saeed",email = "yasrebi.saeed@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "typer (>=0.25.0,<0.26.0)",
    "rich (>=15.0.0,<16.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
